src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/select-in-shadowdom-expected.txt

CONSOLE ERROR: line 18: Uncaught TypeError: host.createShadowRoot is not a function
When selecting from non-anchor Node to anchor node in ShadowDOM, client should not cause page jump.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
